+++ src/josepy/__init__.py
@@ -26,6 +26,9 @@ Originally developed as part of the ACME_ protocol implementation.
 
 """
 
+import sys
+import warnings
+
 # flake8: noqa
 from josepy.b64 import b64decode, b64encode
 from josepy.errors import (
@@ -73,3 +76,10 @@ from josepy.util import (
     ComparableX509,
     ImmutableMap,
 )
+
+if sys.version_info[:2] == (3, 8):
+    warnings.warn(
+        "Python 3.8 support will be dropped in the next scheduled release of "
+        "josepy. Please upgrade your Python version.",
+        DeprecationWarning,
+    )